+++ benchmarks/examples/plot_starrett_et_al_2014_mixtures_fig3.py
@@ -512,8 +512,7 @@ def main() -> None:
     )
     plot_style.close()
 
-    if "agg" not in plt.get_backend().lower():
-        plt.show()
+    plt.show()